Update submodule commits for Services_MS and Permissions_MS; implement permission verification in user and setor controllers

+++ controller/Service/ServicesController.js
@@ -1,12 +1,48 @@
 const service_api = require("../../service/service_api");
 const permissions_api = require("../../service/permissions_api");
+const { verifyPermission } = require("../../utils/verifyPermission");
+
+const SERVICE = 3;
 
 const getAllServices = async (request, reply) => {
   try {
-    let response = await service_api.get("/service");
-    let services = response.data;
+    let user = request.user;
+    let authorized = await verifyPermission(user, SERVICE, request.method);
 
-    reply.status(200).send(services);
+    if (!authorized) {
+      throw {
+        status: 401,
+        message: "You do not have permission to access this resource.",
+      };
+    }
+
+    let serviceResponse = await service_api.get("/service");
+    let services = serviceResponse.data;
+    let permissionsResponse = await permissions_api.get(`/permission/service`);
+    let permissions = permissionsResponse.data;
+
+    let roles = await permissions_api.get(`/roles`);
+    let userServices = [];
+
+    let role_id = roles.data.find((role) => role.name === user.role).id;
+
+    if (!role_id) {
+      throw { status: 500, message: "Role not found" };
+    }
+
+    services.forEach((service) => {
+      let userPermission = permissions.find(
+        (permission) =>
+          permission.role_id === role_id && permission.service_id === service.id
+      );
+      if (userPermission?.read) {
+        userServices = [...userServices, service];
+      }
+    });
+
+
+
+    reply.status(200).send({services: userServices});
   } catch (error) {
     throw error;
   }
@@ -15,8 +51,19 @@ const getAllServices = async (request, reply) => {
 const getService = async (request, reply) => {
   try {
     let id = request.params.id;
+    let user = request.user;
+
+    let authorized = await verifyPermission(user, SERVICE, request.method);
+
+    if (!authorized) {
+      throw {
+        status: 401,
+        message: "You do not have permission to access this resource.",
+      };
+    }
     let response = await service_api.get(`/service/${id}`);
     let services = response.data;
+
 
     reply.status(200).send(services);
   } catch (error) {
@@ -27,6 +74,16 @@ const getService = async (request, reply) => {
 const createService = async (request, reply) => {
   try {
     let service = request.body.service;
+    let user = request.user;
+
+    let authorized = await verifyPermission(user, SERVICE, request.method);
+
+    if (!authorized) {
+      throw {
+        status: 401,
+        message: "You do not have permission to access this resource.",
+      };
+    }
 
     let response = await service_api.post(`/service`, {
       service,
@@ -53,6 +110,18 @@ const updateService = async (request, reply) => {
       let service = request.body.service;
       let permissions = request.body.service.permissions;
       let id = request.params.id;
+
+      let user = request.user;
+
+      let authorized = await verifyPermission(user, SERVICE, request.method);
+
+      if (!authorized) {
+        throw {
+          status: 401,
+          message: "You do not have permission to access this resource.",
+        };
+      }
+
       await service_api.put(`/service/${id}`, {
         service: {
           name: service.name,
@@ -61,8 +130,7 @@ const updateService = async (request, reply) => {
         },
       });
 
-      await service_api.put(`/permissions/service/${id}`, {permissions});
-      console.log(id, { permissions });
+      await permissions_api.put(`/permission/service/${id}`, { permissions });
 
       reply.status(204);
     } catch (error) {
@@ -76,7 +144,19 @@ const updateService = async (request, reply) => {
 const deleteService = async (request, reply) => {
   try {
     let id = request.params.id;
+    let user = request.user;
+
+    let authorized = await verifyPermission(user, SERVICE, request.method);
+
+    if (!authorized) {
+      throw {
+        status: 401,
+        message: "You do not have permission to access this resource.",
+      };
+    }
+
     await service_api.delete(`/service/${id}`);
+    await permissions_api.delete(`/permission/service/${id}`);
 
     reply.status(204);
   } catch (error) {
