+++ controller/Setor/setorController.js
@@ -1,7 +1,19 @@
 const setor_api = require('../../service/setor_api');
+const {verifyPermission} = require('../../utils/verifyPermission')
+
+
+const SERVICE = 2
 
 const getSetores = async (request, reply) => {
   try {
+    let user = request.user
+
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
+
     const response = await setor_api.get('/setor');
 
     let setores = response.data
@@ -14,6 +26,14 @@ const getSetores = async (request, reply) => {
 const getOneSetor = async (request, reply) => {
   try {
     let id = request.params.id;
+    let user = request.user
+
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
+
     const response = await setor_api.get(`/setor/${id}`);
 
     let setor = response.data
@@ -25,6 +45,14 @@ const getOneSetor = async (request, reply) => {
 
 const createSetor = async (request, reply) => {
   try {
+    let user = request.user
+
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
+
     let setor = request.body.setor;
     const response = await setor_api.post('/setor', {setor});
 
@@ -38,6 +66,14 @@ const createSetor = async (request, reply) => {
 const  updateSetor = async (request, reply) => {
   try {
     let id = request.params.id;
+    let user = request.user
+
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
+
     let setor = request.body.setor;
     const response = await setor_api.put(`/setor/${id}`, {setor});
 
@@ -50,7 +86,14 @@ const  updateSetor = async (request, reply) => {
 const deleteSetor = async (request, reply) => {
   try {
     let id = request.params.id;
-    const response = await setor_api.delete(`/setor/${id}`);
+    let user = request.user
+
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
+    await setor_api.delete(`/setor/${id}`);
 
     reply.status(204);
   } catch (error) {
