+++ controller/User/UserController.js
@@ -1,27 +1,30 @@
 const user_api = require("../../service/user_api");
 const permission_api = require("../../service/permissions_api");
+const {verifyPermission} = require('../../utils/verifyPermission')
 
+const SERVICE = 1;
 const cadastrarUser = async (request, reply) => {
   try {
-    let user = request.body.user;
+    let userTarget = request.body.user;
 
     let response = await permission_api.get("/roles");
     let permissions = response.data;
 
     await console.log("permissions: ", permissions);
-    await console.log("user: ", user);
+    await console.log("user: ", userTarget);
 
     const isValid = permissions.some((permission) => {
-      return permission.name === user.role;
+      return permission.name === userTarget.role;
     });
 
     if (!isValid) {
       throw { status: 403, message: "Role not found" };
     }
+
     console.log("isvalid: ", isValid);
     await user_api.post("/user", {
       user: {
-        ...user,
+        ...userTarget,
       },
     });
 
@@ -34,10 +37,18 @@ const cadastrarUser = async (request, reply) => {
 const getOneUser = async (request, reply) => {
   try {
     let id = request.params.id;
-    let response = await user_api.get(`/user/${id}`);
-    let user = response.data;
+    let user = request.user
 
-    reply.status(200).send(user);
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
+
+    let response = await user_api.get(`/user/${id}`);
+    let userTarget = response.data;
+
+    reply.status(200).send(userTarget);
   } catch (error) {
     throw error;
   }
@@ -45,10 +56,18 @@ const getOneUser = async (request, reply) => {
 
 const getAllUser = async (request, reply) => {
   try {
-    let response = await user_api.get(`/user`);
-    let users = response.data;
+    let user = request.user
 
-    reply.status(200).send(users);
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
+
+    let response = await user_api.get(`/user`);
+    let usersTarget = response.data;
+
+    reply.status(200).send(usersTarget);
   } catch (error) {
     throw error;
   }
@@ -57,11 +76,18 @@ const getAllUser = async (request, reply) => {
 const atualizarUser = async (request, reply) => {
   try {
     let id = request.params.id;
-    let user = request.body.user;
+    let userTarget = request.body.user;
+    let user = request.user
+
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
 
     await user_api.put(`/user/${id}`, {
       user: {
-        ...user,
+        ...userTarget,
       },
     });
     reply.status(200).send("usuário atualizado com sucesso");
@@ -73,6 +99,14 @@ const atualizarUser = async (request, reply) => {
 const deletarUser = async (request, reply) => {
   try {
     let id = request.params.id;
+    let user = request.user
+
+    let authorized = await verifyPermission(user, SERVICE, request.method) 
+
+    if(!authorized){
+      throw { status: 401, message: "You do not have permission to access this resource." }
+    }
+
     await user_api.delete(`/user/${id}`);
 
     reply.status(200).send("Usuário excluido com sucesso");
